+++ qgis_plugin_manager/__main__.py
@@ -24,6 +24,7 @@ from qgis_plugin_manager.remote import (
 from qgis_plugin_manager.utils import (
     PluginManagerError,
     get_semver_version,
+    get_semver_version_str,
     install_epilog,
     print_json,
     print_table,
@@ -354,7 +355,6 @@ def install_plugin(args: Namespace):
         plugin_name = parameter[0]
 
         plugin_info = plugins.plugin_info(plugin_name)
-
         if len(parameter) >= 2:
             plugin_version = parameter[1]
             if not plugin_version:
@@ -371,7 +371,7 @@ def install_plugin(args: Namespace):
                 if latest and latest.version == plugin_info.version:
                     echo.alert(f"\t{plugin_name}=={plugin_info.version} is already at latest version")
                     continue
-            elif plugin_version is None and plugin_info.version == plugin_version:
+            elif plugin_info.version == get_semver_version_str(plugin_version):
                 echo.alert(f"\t{plugin_name}=={plugin_version} already installed")
                 continue
             elif plugin_version is None: